feat(gfx-vulkan): Refactor vulkan instance creation
Move this into a core function so that instances are created in a standard way
